+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 
 ### Changed
 
+- Make `SurfaceMesh::bounds()` a free function `pmp::bounds()`.
 - Remove `SurfaceMesh::edge_length()`.
 - Remove object properties.
 - Remove support for PLY format, see 1ec912b for why.